Restored the commit log. Dealing with gradle issues.

+++ not_commented/app/src/main/java/com/gps/conversorgalaxy/MainActivity.java
@@ -201,12 +201,12 @@ public class MainActivity extends AppCompatActivity {
                 } else if (selectedConversion.equals("normal") && ((8 < i) && (i < 49))) {
                     modifier = 4;
                 } else {
-                    modifier = (int) -i;
+                    modifier = -i;
                 }
                 break;
             case " - Dimension: ":
                 if (((16 < i) && (i < 33)) && selectedConversion.equals("galaxy"))
-                    modifier = (int) -i;
+                    modifier = -i;
                 else if ((16 < i) && (i < 529))
                     modifier = -16;
                 else
